# docs/api/dgraph-openapi.yaml
openapi: 3.0.3
info:
  title: DGraph Service API
  description: |
    Off-chain and social data. All social interactions (posts, feed, interactions),
    prediction markets (list, bets, payouts), governance variables, subscription status,
    location preferences and spots, ad campaigns/impressions/analytics, and safety
    run through this service (read/write from Dgraph or in-memory fallback).
  version: 1.0.0
servers:
  - url: http://localhost:3003
    description: Local default
  - url: https://gql.dlux.io
    description: Test/production
...
paths:
...
  /location/preferences:
    get:
      tags: [location]
      summary: Get location preferences
      parameters:
        - name: user
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: preferences
    post:
      tags: [location]
      summary: Update location preferences
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user: { type: string }
                enabled: { type: boolean }
                precision: { type: string }
                subscribedSpots: { type: array, items: { type: string } }
      responses:
        '200':
          description: Updated
...
  /location/spots/subscribe:
    post:
      tags: [location]
      summary: Subscribe to spot
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user: { type: string }
                spotId: { type: string }
      responses:
        '200':
          description: Subscribed
...